+++ week07/ExerciseTracking/Running.cs
@@ -1,23 +1,23 @@
 public class Running : Activity
 {
-    private float _distance;
+    private double _distance;
 
-    public Running(string date, float length, float distance) : base (date, length)
+    public Running(string date, double length, double distance) : base (date, length)
     {
         _distance = distance;
     } 
 
-    public override float GetDistance()
+    public override double GetDistance()
     {
         return _distance;
     }
 
-    public override float GetSpeed()
+    public override double GetSpeed()
     {
         return _distance / GetLength() * 60;
     }
 
-    public override float GetPace()
+    public override double GetPace()
     {
         return GetLength() / _distance;
     }
