+++ week07/ExerciseTracking/Swimming.cs
@@ -2,22 +2,22 @@ public class Swimming : Activity
 {
     private int _laps;
 
-    public Swimming(string date, float length, int laps) : base(date, length)
+    public Swimming(string date, double length, int laps) : base(date, length)
     {
         _laps = laps;
     }
 
-    public override float GetDistance()
+    public override double GetDistance()
     {
         return _laps * 50 / 1000;
     }
 
-    public override float GetSpeed()
+    public override double GetSpeed()
     {
         return GetDistance() / GetLength() * 60;
     }
 
-    public override float GetPace()
+    public override double GetPace()
     {
         return GetLength() / GetDistance();
     }
